admin dashboard - tag/category delete

+++ app/Http/Controllers/TagController.php
@@ -58,6 +58,10 @@ class TagController extends Controller
     public function store(TagRequest $request)
     {
         try {
+            if ($request->tag_id == '1') {
+                return $this->errorFailMessage('Can not modify default resource!');
+            }
+
             $result = $this->tagService->ajaxStore($request);
 
             if ($result) {
@@ -139,34 +143,18 @@ class TagController extends Controller
     public function destroy($id)
     {
         try {
+            if ($id == '1') {
+                return $this->errorFailMessage('Can not delete default resource!');
+            }
+
             $tag = $this->tagService->findById($id);
             $result = $this->tagService->destroy($tag);
+
             if ($result) {
                 return response()->json(['success' => 'Tag deleted successfully']);
-            }
-            if (false == $result) {
-                return response()->json([
-                    'status' => 202,
-                    'errors' => ['Failed!', 'Can not delete default resource!'],
-                ]);
             }
 
             return $this->errorFailMessage();
-        } catch (\Exception $e) {
-            return $this->errorMessage();
-            $tag = Tag::findById($id);
-            // update new value for each post before delete Tag
-            $tag->posts()->whereTagId($id)->update(['tag_id' => 1]);
-            $tag->posts()->detach();
-
-            $result = $tag->delete();
-
-            $result = $this->tagService->destroy($id);
-            if ($result) {
-                return response()->json(['success' => 'Tag deleted successfully']);
-            }
-
-            return response()->json($result);
         } catch (\Exception $e) {
             return response()->json($e->getMessage());
         }
@@ -241,7 +229,6 @@ class TagController extends Controller
         // Toastr::error('Something went wrong!', 'Error');
     }
 
-
     /**
      * Display validation errors of request.
      */
@@ -267,15 +254,14 @@ class TagController extends Controller
     /**
      * Display failed errors of request.
      */
-    protected function errorFailMessage()
+    protected function errorFailMessage($msg = null)
     {
-        $msg = [
+        $message = [
             'status' => 500,
-            'errors' => ['Failed!', 'Unknown error!'],
-
+            'errors' => ['Failed!', $msg],
             'success' => false,
         ];
 
-        return response()->json($msg);
+        return response()->json($message);
     }
 }
